refactor(Assistant): Extract some helpers from provider
