+++ src/assistant/Conversations/ChatConversation.jsx
@@ -3,7 +3,8 @@ import React, { useRef, useEffect } from 'react'
 import { useQuery, isQueryLoading } from 'cozy-client'
 
 import { buildChatConversationQueryById } from '../queries'
-import { useAssistant, getInstantMessage } from '../AssistantProvider'
+import { useAssistant } from '../AssistantProvider'
+import { getInstantMessage } from '../helpers'
 import ChatUserItem from './ChatUserItem'
 import ChatAssistantItem from './ChatAssistantItem'
 import ChatRealtimeAnswer from './ChatRealtimeAnswer'
