+++ src/assistant/Views/SearchDialog.jsx
@@ -5,7 +5,8 @@ import flag from 'cozy-flags'
 import { FixedDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
 
 import SearchProvider from '../Search/SearchProvider'
-import { makeConversationId, useAssistant } from '../AssistantProvider'
+import { useAssistant } from '../AssistantProvider'
+import { makeConversationId } from '../helpers'
 import ResultMenuContent from '../ResultMenu/ResultMenuContent'
 import { useSearch } from '../Search/SearchProvider'
 import SearchBar from '../Search/SearchBar'
